moved tests to  RedisQuesProcessorTest.java
